requirements.txt
streamlit==1.31.0
pandas==2.2.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
redis==5.0.1
google-generativeai==0.3.1
numpy==1.26.3
plotly==5.18.0 